Module Tezos_009_PsFLoren_test_helpers__Block.Forge

val contents : ?⁠proof_of_work_nonce:Stdlib.Bytes.t -> ?⁠priority:int -> ?⁠seed_nonce_hash:Tezos_raw_protocol_009_PsFLoren.Nonce_hash.t -> unit -> Tezos_raw_protocol_009_PsFLoren.Alpha_context.Block_header.contents

type header

val forge_header : ?⁠policy:baker_policy -> ?⁠timestamp:Tezos_raw_protocol_009_PsFLoren.Alpha_context.Timestamp.time -> ?⁠operations:Tezos_raw_protocol_009_PsFLoren.Alpha_context.Operation.packed list -> t -> header Tezos_base__TzPervasives.tzresult Lwt.t

Forges a correct header following the policy. The header can then be modified and applied with apply.

val set_seed_nonce_hash : Tezos_raw_protocol_009_PsFLoren.Nonce_hash.t option -> header -> header

Sets uniquely seed_nonce_hash of a header

val set_baker : Tezos_raw_protocol_009_PsFLoren.Alpha_context.public_key_hash -> header -> header

Sets the baker that will sign the header to an arbitrary pkh

val sign_header : header -> Tezos_raw_protocol_009_PsFLoren.Alpha_context.Block_header.block_header Tezos_base__TzPervasives.tzresult Lwt.t

Signs the header with the key of the baker configured in the header. The header can no longer be modified, only applied.